Added Delete Feature

+++ Application/modules/fileHandling.py
@@ -1,13 +1,12 @@
 
 class FILE():
-    def __init__(self,item,details):
-        self._item = item
-        self._details = details
-        self._file = open(self._details["path"],"r+")
+    def __init__(self):
+        pass
     
     def openFile(self,item,details):
         self.closeFile() # In case a some previous file was open
 
+        self._item = item
         self._name = item.text(0)
         self._details = details
         self._file = open(self._details["path"],"r+")
@@ -22,10 +21,12 @@ class FILE():
         return self._name
     
     def saveFile(self,text):
+        if self._file.closed:
+            return
         self._file.seek(0)
         self._file.truncate()
         self._file.write(text)
         # self._file.flush()
 
 
-currentNote = FILE("MarkdownGuide",{"path" : "../Application/notes/note4.txt"})
+currentNote = FILE()
